Update some code formatting and comments

+++ virtool/account/api.py
@@ -16,10 +16,7 @@ from virtool.api.response import bad_request, json_response, no_content, not_fou
 
 #: A MongoDB projection to use when returning API key documents to clients. The key should never be sent to client after
 #: its creation.
-API_KEY_PROJECTION = {
-    "_id": False,
-    "user": False
-}
+API_KEY_PROJECTION = {"_id": False, "user": False}
 
 #: A :class:`aiohttp.web.RouteTableDef` for account API routes.
 routes = virtool.http.routes.Routes()
@@ -31,25 +28,22 @@ async def get(req):
     Get complete user document.
 
     """
-    document = await virtool.account.db.get_document(req.app["db"], req["client"].user_id)
+    document = await virtool.account.db.get(req.app["db"], req["client"].user_id)
     return json_response(virtool.utils.base_processor(document))
 
 
-@routes.patch("/api/account", schema={
-    "email": {
-        "type": "string",
-        "coerce": virtool.validators.strip,
-        "regex": r"^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$"
+@routes.patch(
+    "/api/account",
+    schema={
+        "email": {
+            "type": "string",
+            "coerce": virtool.validators.strip,
+            "regex": r"^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$",
+        },
+        "old_password": {"type": "string"},
+        "password": {"type": "string", "dependencies": "old_password"},
     },
-    "old_password": {
-        "type": "string"
-    },
-    "password": {
-        "type": "string",
-        "dependencies": "old_password"
-    }
-
-})
+)
 async def edit(req):
     """
     Edit the user account.
@@ -70,7 +64,9 @@ async def edit(req):
         if error:
             return bad_request(error)
 
-        if not await virtool.users.db.validate_credentials(db, user_id, old_password or ""):
+        if not await virtool.users.db.validate_credentials(
+            db, user_id, old_password or ""
+        ):
             return bad_request("Invalid credentials")
 
         update = virtool.account.db.compose_password_update(password)
@@ -79,11 +75,11 @@ async def edit(req):
         update["email"] = data["email"]
 
     if update:
-        document = await db.users.find_one_and_update({"_id": user_id}, {
-            "$set": update
-        }, projection=virtool.account.db.PROJECTION)
+        document = await db.users.find_one_and_update(
+            {"_id": user_id}, {"$set": update}, projection=virtool.account.db.PROJECTION
+        )
     else:
-        document = await virtool.account.db.get_document(db, user_id)
+        document = await virtool.account.db.get(db, user_id)
 
     return json_response(virtool.utils.base_processor(document))
 
@@ -95,29 +91,24 @@ async def get_settings(req):
 
     """
     account_settings = await virtool.db.utils.get_one_field(
-        req.app["db"].users,
-        "settings",
-        req["client"].user_id
+        req.app["db"].users, "settings", req["client"].user_id
     )
 
     return json_response(account_settings)
 
 
-@routes.patch("/api/account/settings", schema={
-    "show_ids": {
-        "type": "boolean",
-        "required": False
+@routes.patch(
+    "/api/account/settings",
+    schema={
+        "show_ids": {"type": "boolean", "required": False},
+        "skip_quick_analyze_dialog": {"type": "boolean", "required": False},
+        "quick_analyze_workflow": {
+            "type": "string",
+            "allowed": virtool.analyses.utils.WORKFLOW_NAMES,
+            "required": False,
+        },
     },
-    "skip_quick_analyze_dialog": {
-        "type": "boolean",
-        "required": False
-    },
-    "quick_analyze_workflow": {
-        "type": "string",
-        "allowed": virtool.analyses.utils.WORKFLOW_NAMES,
-        "required": False
-    }
-})
+)
 async def update_settings(req):
     """
     Update account settings.
@@ -128,16 +119,13 @@ async def update_settings(req):
 
     user_id = req["client"].user_id
 
-    settings_from_db = await virtool.db.utils.get_one_field(db.users, "settings", user_id)
+    settings_from_db = await virtool.db.utils.get_one_field(
+        db.users, "settings", user_id
+    )
 
-    settings = {
-        **settings_from_db,
-        **data
-    }
+    settings = {**settings_from_db, **data}
 
-    await db.users.update_one({"_id": user_id}, {
-        "$set": settings
-    })
+    await db.users.update_one({"_id": user_id}, {"$set": settings})
 
     return json_response(settings)
 
@@ -159,7 +147,9 @@ async def get_api_key(req):
     user_id = req["client"].user_id
     key_id = req.match_info["key_id"]
 
-    document = await db.keys.find_one({"id": key_id, "user.id": user_id}, API_KEY_PROJECTION)
+    document = await db.keys.find_one(
+        {"id": key_id, "user.id": user_id}, API_KEY_PROJECTION
+    )
 
     if document is None:
         return not_found()
@@ -167,19 +157,22 @@ async def get_api_key(req):
     return json_response(document, status=200)
 
 
-@routes.post("/api/account/keys", schema={
-    "name": {
-        "type": "string",
-        "coerce": virtool.validators.strip,
-        "empty": False,
-        "required": True
+@routes.post(
+    "/api/account/keys",
+    schema={
+        "name": {
+            "type": "string",
+            "coerce": virtool.validators.strip,
+            "empty": False,
+            "required": True,
+        },
+        "permissions": {
+            "type": "dict",
+            "default": {},
+            "validator": virtool.validators.is_permission_dict,
+        },
     },
-    "permissions": {
-        "type": "dict",
-        "default": {},
-        "validator": virtool.validators.is_permission_dict
-    }
-})
+)
 async def create_api_key(req):
     """
     Create a new API key. The new key value is returned in the response. This is the only response from the server that
@@ -192,26 +185,24 @@ async def create_api_key(req):
     user_id = req["client"].user_id
 
     document = await virtool.account.db.create_api_key(
-        db,
-        data["name"],
-        data["permissions"],
-        user_id
+        db, data["name"], data["permissions"], user_id
     )
 
-    headers = {
-        "Location": f"/api/account/keys/{document['id']}"
-    }
+    headers = {"Location": f"/api/account/keys/{document['id']}"}
 
     return json_response(document, headers=headers, status=201)
 
 
-@routes.patch("/api/account/keys/{key_id}", schema={
-    "permissions": {
-        "type": "dict",
-        "validator": virtool.validators.is_permission_dict,
-        "required": True
-    }
-})
+@routes.patch(
+    "/api/account/keys/{key_id}",
+    schema={
+        "permissions": {
+            "type": "dict",
+            "validator": virtool.validators.is_permission_dict,
+            "required": True,
+        }
+    },
+)
 async def update_api_key(req):
     """
     Change the permissions for an existing API key.
@@ -231,21 +222,21 @@ async def update_api_key(req):
 
     # The permissions currently assigned to the API key.
     permissions = await virtool.db.utils.get_one_field(
-        db.keys,
-        "permissions",
-        {"id": key_id, "user.id": user_id}
+        db.keys, "permissions", {"id": key_id, "user.id": user_id}
     )
 
     permissions.update(data["permissions"])
 
     if not user["administrator"]:
-        permissions = virtool.users.utils.limit_permissions(permissions, user["permissions"])
+        permissions = virtool.users.utils.limit_permissions(
+            permissions, user["permissions"]
+        )
 
-    document = await db.keys.find_one_and_update({"id": key_id}, {
-        "$set": {
-            "permissions": permissions
-        }
-    }, projection=API_KEY_PROJECTION)
+    document = await db.keys.find_one_and_update(
+        {"id": key_id},
+        {"$set": {"permissions": permissions}},
+        projection=API_KEY_PROJECTION,
+    )
 
     return json_response(document)
 
@@ -256,10 +247,7 @@ async def remove_api_key(req):
     user_id = req["client"].user_id
     key_id = req.match_info["key_id"]
 
-    delete_result = await db.keys.delete_one({
-        "id": key_id,
-        "user.id": user_id
-    })
+    delete_result = await db.keys.delete_one({"id": key_id, "user.id": user_id})
 
     if delete_result.deleted_count == 0:
         return not_found()
@@ -277,22 +265,15 @@ async def remove_all_api_keys(req):
     return no_content()
 
 
-@routes.post("/api/account/login", public=True, schema={
-    "username": {
-        "type": "string",
-        "empty": False,
-        "required": True
+@routes.post(
+    "/api/account/login",
+    public=True,
+    schema={
+        "username": {"type": "string", "empty": False, "required": True},
+        "password": {"type": "string", "empty": False, "required": True},
+        "remember": {"type": "boolean", "default": False},
     },
-    "password": {
-        "type": "string",
-        "empty": False,
-        "required": True
-    },
-    "remember": {
-        "type": "boolean",
-        "default": False
-    }
-})
+)
 async def login(req):
     """
     Create a new session for the user with `username`.
@@ -316,17 +297,18 @@ async def login(req):
     # If the user's password needs to be reset, redirect to the reset page without authorizing the session. A one-time
     # reset code is generated and added to the query string.
     if await virtool.db.utils.get_one_field(db.users, "force_reset", user_id):
-        return json_response({
-            "reset": True,
-            "reset_code": await virtool.users.sessions.create_reset_code(db, session_id, user_id, remember)
-        }, status=200)
+        return json_response(
+            {
+                "reset": True,
+                "reset_code": await virtool.users.sessions.create_reset_code(
+                    db, session_id, user_id, remember
+                ),
+            },
+            status=200,
+        )
 
     session, token = await virtool.users.sessions.replace_session(
-        db,
-        session_id,
-        virtool.http.auth.get_ip(req),
-        user_id,
-        remember
+        db, session_id, virtool.http.auth.get_ip(req), user_id, remember
     )
 
     resp = json_response({"reset": False}, status=201)
@@ -347,9 +329,7 @@ async def logout(req):
     old_session_id = req.cookies.get("session_id")
 
     session, _ = await virtool.users.sessions.replace_session(
-        db,
-        old_session_id,
-        virtool.http.auth.get_ip(req)
+        db, old_session_id, virtool.http.auth.get_ip(req)
     )
 
     resp = aiohttp.web.Response(status=200)
@@ -360,22 +340,17 @@ async def logout(req):
     return resp
 
 
-@routes.post("/api/account/reset", public=True, schema={
-    "password": {
-        "type": "string",
-        "required": True
+@routes.post(
+    "/api/account/reset",
+    public=True,
+    schema={
+        "password": {"type": "string", "required": True},
+        "reset_code": {"type": "string", "required": True},
     },
-    "reset_code": {
-        "type": "string",
-        "required": True
-    }
-})
+)
 async def reset(req):
     """
     Handles `POST` requests for resetting the password for a session user.
-
-    :param req: the request to handle
-    :return: a response
 
     """
     db = req.app["db"]
@@ -388,18 +363,27 @@ async def reset(req):
 
     error = await virtool.users.checks.check_password_length(req)
 
-    if not session.get("reset_code") or not session.get("reset_user_id") or reset_code != session.get("reset_code"):
+    if (
+        not session.get("reset_code")
+        or not session.get("reset_user_id")
+        or reset_code != session.get("reset_code")
+    ):
         error = "Invalid reset code"
 
     user_id = session["reset_user_id"]
 
     if error:
-        return json_response({
-            "error": error,
-            "reset_code": await virtool.users.sessions.create_reset_code(db, session_id, user_id=user_id)
-        }, status=400)
+        return json_response(
+            {
+                "error": error,
+                "reset_code": await virtool.users.sessions.create_reset_code(
+                    db, session_id, user_id=user_id
+                ),
+            },
+            status=400,
+        )
 
-    # Update the user password and disable the `force_reset`.
+    # Update the user password and disable `force_reset`.
     await virtool.users.db.edit(db, user_id, force_reset=False, password=password)
 
     new_session, token = await virtool.users.sessions.replace_session(
@@ -407,15 +391,12 @@ async def reset(req):
         session_id,
         virtool.http.auth.get_ip(req),
         user_id,
-        remember=session.get("reset_remember", False)
+        remember=session.get("reset_remember", False),
     )
 
     req["client"].authorize(new_session, is_api=False)
 
-    resp = json_response({
-        "login": False,
-        "reset": False
-    }, status=200)
+    resp = json_response({"login": False, "reset": False}, status=200)
 
     virtool.http.utils.set_session_id_cookie(resp, new_session["_id"])
     virtool.http.utils.set_session_token_cookie(resp, token)
